requirements.txt
absl-py >= 1.2.0
google-api-python-client
google-auth

